chore(clippy): Simplify minor stuff found by clippy

+++ src/cargo/ops/registry.rs
@@ -604,15 +604,15 @@ pub fn search(
     }
 
     let search_max_limit = 100;
-    if total_crates > u32::from(limit) && limit < search_max_limit {
+    if total_crates > limit && limit < search_max_limit {
         println!(
             "... and {} crates more (use --limit N to see more)",
-            total_crates - u32::from(limit)
+            total_crates - limit
         );
-    } else if total_crates > u32::from(limit) && limit >= search_max_limit {
+    } else if total_crates > limit && limit >= search_max_limit {
         println!(
             "... and {} crates more (go to http://crates.io/search?q={} to see more)",
-            total_crates - u32::from(limit),
+            total_crates - limit,
             percent_encode(query.as_bytes(), QUERY_ENCODE_SET)
         );
     }
